Move some methods into more appropriate files
