+++ pkg/cloud/aws/actuators/machine/security_groups.go
@@ -109,70 +109,3 @@ func (a *Actuator) securityGroupsChanged(annotation map[string]interface{}, addi
 
 	return changed, res
 }
-
-// tagsChanged determines which tags to delete and which to add.
-func (a *Actuator) tagsChanged(annotation map[string]interface{}, src map[string]string) (bool, map[string]string, map[string]string, map[string]interface{}) {
-	// Bool tracking if we found any changed state.
-	changed := false
-
-	// Tracking for created/updated
-	created := map[string]string{}
-
-	// Tracking for tags that were deleted.
-	deleted := map[string]string{}
-
-	// The new annotation that we need to set if anything is created/updated.
-	newAnnotation := map[string]interface{}{}
-
-	// Loop over annotation, checking if entries are in src.
-	// If an entry is present in annotation but not src, it has been deleted
-	// since last time. We flag this in the deleted map.
-	for t, v := range annotation {
-		_, ok := src[t]
-
-		// Entry isn't in src, it has been deleted.
-		if !ok {
-			// Cast v to a string here. This should be fine, tags are always
-			// strings.
-			deleted[t] = v.(string)
-			changed = true
-		}
-	}
-
-	// Loop over src, checking for entries in annotation.
-	//
-	// If an entry is in src, but not annotation, it has been created since
-	// last time.
-	//
-	// If an entry is in both src and annotation, we compare their values, if
-	// the value in src differs from that in annotation, the tag has been
-	// updated since last time.
-	for t, v := range src {
-		av, ok := annotation[t]
-
-		// Entries in the src always need to be noted in the newAnnotation. We
-		// know they're going to be created or updated.
-		newAnnotation[t] = v
-
-		// Entry isn't in annotation, it's new.
-		if !ok {
-			created[t] = v
-			newAnnotation[t] = v
-			changed = true
-			continue
-		}
-
-		// Entry is in annotation, has the value changed?
-		if v != av {
-			created[t] = v
-			changed = true
-		}
-
-		// Entry existed in both src and annotation, and their values were
-		// equal. Nothing to do.
-	}
-
-	// We made it through the loop, and everything that was in src, was also
-	// in dst. Nothing changed.
-	return changed, created, deleted, newAnnotation
-}
